week2/day5/restaurant_menu/requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.7